trying fix 6 issues
