+++ api/case_dashboard/handlers.js
@@ -1,9 +1,7 @@
 const { requestIfSame } = require('../../helpers/request')
 const replyHelper = require('../helpers')
-const json2xls = require('json2xls')
-const moment = require('moment')
-const fs = require('fs')
 const { ROLE } = require('../../helpers/constant')
+const { generateExcell } = require('../../helpers/export')
 
 const countSectionTop = (server) => {
   return async (request, reply) => {
@@ -107,16 +105,6 @@ const mapingCriteria = async (result, role) => {
         'Meninggal': decease,
       }
   ))
-}
-
-const generateExcell = async (data, title, fullName, reply,) => {
-  const jsonXls = json2xls(data);
-  const fileName = `${title}-${fullName}-${moment().format("YYYY-MM-DD-HH-mm")}.xlsx`
-  fs.writeFileSync(fileName, jsonXls, 'binary');
-  const xlsx = fs.readFileSync(fileName)
-  reply(xlsx)
-    .header('Content-Disposition', 'attachment; filename=' + fileName);
-  return fs.unlinkSync(fileName)
 }
 
 module.exports = {
